+++ pdf/CHANGELOG.md
@@ -6,6 +6,7 @@
 * Fix RichText Widget word spacing
 * Improve Theme and TextStyle
 * Implement properly RichText.softWrap
+* Set a proper value to context.pagesCount
 
 # 1.3.7
 * Add Pdf Creation date